+++ src/integration/java/io/pinecone/integration/dataplane/PineconeClientLiveIntegTest.java
@@ -1,59 +1,39 @@
-package io.pinecone;
+package io.pinecone.integration.dataplane;
 
 import com.google.common.primitives.Floats;
 import com.google.protobuf.Struct;
 import com.google.protobuf.Value;
+import io.pinecone.PineconeConnection;
 import io.pinecone.helpers.RandomStringBuilder;
-import io.pinecone.model.IndexMeta;
 import io.pinecone.proto.*;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import static io.pinecone.helpers.IndexManager.createIndexIfNotExistsDataPlane;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.notNullValue;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PineconeClientLiveIntegTest {
 
-    public String indexName = "integ-test-sanity";
-
     private static final Logger logger = LoggerFactory.getLogger(PineconeClientLiveIntegTest.class);
+    private static VectorServiceGrpc.VectorServiceBlockingStub blockingStub;
 
-    private PineconeClient dataPlaneClient;
-    private PineconeIndexOperationClient controlPlaneClient;
-
-    @BeforeEach
-    public void setUp() {
-        PineconeClientConfig config = new PineconeClientConfig()
-                .withApiKey(System.getenv("PINECONE_API_KEY"))
-                .withEnvironment(System.getenv("PINECONE_ENVIRONMENT"))
-                .withServerSideTimeoutSec(10);
-
-        controlPlaneClient = new PineconeIndexOperationClient(config);
-        dataPlaneClient = new PineconeClient(config);
+    @BeforeAll
+    public static void defineConfig() throws IOException, InterruptedException {
+        PineconeConnection connection = createIndexIfNotExistsDataPlane(3);
+        blockingStub = connection.getBlockingStub();
     }
 
     @Test
-    public void checkIndexSetup() throws Exception {
-        IndexMeta indexMeta = controlPlaneClient.describeIndex(indexName);
-        assertEquals(3, indexMeta.getDatabase().getDimension());
-    }
-
-    @Test
-    public void sanity() throws Exception {
-        IndexMeta indexMeta = controlPlaneClient.describeIndex(indexName);
-        String host = indexMeta.getStatus().getHost();
-        PineconeConnection conn = dataPlaneClient.connect(
-                new PineconeConnectionConfig()
-                        .withConnectionUrl("https://" + host));
-
+    public void sanity() {
         String namespace = RandomStringBuilder.build("ns", 8);
 
         // upsert
@@ -76,7 +56,7 @@ public class PineconeClientLiveIntegTest {
                 .setNamespace(namespace)
                 .build();
 
-        UpsertResponse upsertResponse = conn.getBlockingStub().upsert(request);
+        UpsertResponse upsertResponse = blockingStub.upsert(request);
         logger.info("Put " + upsertResponse.getUpsertedCount() + " vectors into the index");
         assert (upsertResponse.getUpsertedCount() == 3);
 
@@ -100,14 +80,14 @@ public class PineconeClientLiveIntegTest {
                 .addAllVectors(hybridVectors)
                 .setNamespace(namespace)
                 .build();
-        UpsertResponse hybridResponse = conn.getBlockingStub().upsert(hybridRequest);
+        UpsertResponse hybridResponse = blockingStub.upsert(hybridRequest);
         logger.info("Put " + hybridResponse.getUpsertedCount() + " vectors into the index");
         assert (hybridResponse.getUpsertedCount() == 3);
 
         // fetch
         List<String> ids = Arrays.asList("v1", "v2");
         FetchRequest fetchRequest = FetchRequest.newBuilder().addAllIds(ids).setNamespace(namespace).build();
-        FetchResponse fetchResponse = conn.getBlockingStub().fetch(fetchRequest);
+        FetchResponse fetchResponse = blockingStub.fetch(fetchRequest);
         assert (fetchResponse.containsVectors("v1"));
 
         // Updates vector v1's values to 10.0, 11.0, and 12.0 from 1.0, 2.0, and 3.0
@@ -116,9 +96,9 @@ public class PineconeClientLiveIntegTest {
                 .setNamespace(namespace)
                 .addAllValues(Floats.asList(10F, 11F, 12F))
                 .build();
-        conn.getBlockingStub().update(updateRequest);
+        blockingStub.update(updateRequest);
         fetchRequest = FetchRequest.newBuilder().addIds("v1").setNamespace(namespace).build();
-        conn.getBlockingStub().fetch(fetchRequest);
+        blockingStub.fetch(fetchRequest);
 
         // DEPRECATED: all methods related to queries in QueryVector
         // Use methods related to Vector. Example: addVector, addAllVector, etc.
@@ -145,7 +125,7 @@ public class PineconeClientLiveIntegTest {
                 .setIncludeMetadata(true)
                 .build();
 
-        QueryResponse queryResponse = conn.getBlockingStub().query(batchQueryRequest);
+        QueryResponse queryResponse = blockingStub.query(batchQueryRequest);
         assertThat(queryResponse, notNullValue());
         assertThat(queryResponse.getResultsList(), notNullValue());
         assertThat(queryResponse.getResultsCount(), equalTo(1));
@@ -159,7 +139,7 @@ public class PineconeClientLiveIntegTest {
                 .build();
 
         // When querying using a single vector, we get matches instead of results
-        queryResponse = conn.getBlockingStub().query(queryRequest);
+        queryResponse = blockingStub.query(queryRequest);
         assertThat(queryResponse, notNullValue());
         assertThat(queryResponse.getMatchesList(), notNullValue());
         assertThat(queryResponse.getMatchesCount(), equalTo(2));
@@ -172,7 +152,7 @@ public class PineconeClientLiveIntegTest {
                 .setIncludeMetadata(true)
                 .build();
 
-        queryResponse = conn.getBlockingStub().query(queryByIdRequest);
+        queryResponse = blockingStub.query(queryByIdRequest);
         assertThat(queryResponse, notNullValue());
         assertThat(queryResponse.getMatchesList(), notNullValue());
         assertThat(queryResponse.getMatchesCount(), equalTo(2));
@@ -185,9 +165,9 @@ public class PineconeClientLiveIntegTest {
                 .setDeleteAll(false)
                 .build();
 
-        conn.getBlockingStub().delete(deleteRequest);
+        blockingStub.delete(deleteRequest);
         fetchRequest = FetchRequest.newBuilder().addAllIds(ids).setNamespace(namespace).build();
-        fetchResponse = conn.getBlockingStub().fetch(fetchRequest);
+        fetchResponse = blockingStub.fetch(fetchRequest);
         assert (fetchResponse.getVectorsCount() == ids.size() - 1);
 
         // Clear out the test
@@ -196,9 +176,9 @@ public class PineconeClientLiveIntegTest {
                 .setDeleteAll(true)
                 .build();
 
-        conn.getBlockingStub().delete(deleteAllRequest);
+        blockingStub.delete(deleteAllRequest);
         fetchRequest = FetchRequest.newBuilder().addAllIds(ids).setNamespace(namespace).build();
-        fetchResponse = conn.getBlockingStub().fetch(fetchRequest);
+        fetchResponse = blockingStub.fetch(fetchRequest);
         assert (fetchResponse.getVectorsCount() == 0);
     }
 }